chore: Rename color-console.js to helper.js and add askQuestion as exported fn
